// src/app/models/verify-student-application.ts
export class VerifyStudentApplication {
    tempStudentId: any;
    fullName: string = '';
    email: string = '';
    nationality: string = '';
    telephone: string = '';
    nic: string = '';
    gender: string = '';
    createdDate: any;

    qualificationName: string = '';
    qualificationTypeName: string = '';
    school: string = '';
    overallGrade: string = '';
    effectiveDate: any;

    courseId: any;
    courseTypeId: any;
    batchId: any;
    courseProviderId: any;
    smsAccountId: any;
    paymentStatus: string = '';
    amount: string = '';
    paymentDate: any;
    courseOfferId: any;

    branchId: any;
    offerName: string = '';
    username: any;
    status: any;
}